requirements.txt
matplotlib==3.1.1
~cipy==1.3.1
tensorflow==1.15.0
numpy==1.16.5
altair==4.1.0
opencv_python==4.2.0.32
streamlit==0.78.0
librosa==0.8.0
pandas==0.25.1
python_speech_features==0.6
Pillow==8.2.0
scipy==1.6.2
